Split typechecker semantic guard tests
