+++ tests/docs_truth/repo_hygiene/file_size.rs
@@ -11,4 +11,5 @@ mod resolver_phase2_splits;
 mod resolver_validation_splits;
 mod resolver_validation_test_splits;
 mod thresholds;
+mod typechecker_semantic_splits;
 mod typechecker_test_splits;
