+++ tests/docs_truth/repo_hygiene/file_size/typechecker_test_splits.rs
@@ -167,87 +167,3 @@ fn resolver_required_stale_diagnostic_tests_live_in_focused_helper() {
         "required_diagnostics.rs should include the focused stale_requires module"
     );
 }
-
-#[test]
-fn declaration_validation_precollection_tasks_live_in_focused_helper() {
-    let tasks = read("src/typechecker/tests/declaration_validation/tasks.rs");
-    let precollection = read("src/typechecker/tests/declaration_validation/precollection_tasks.rs");
-    let module = read("src/typechecker/tests/declaration_validation.rs");
-
-    assert!(
-        tasks.lines().count() < 240,
-        "tasks.rs should stay focused on declaration semantic validation tasks"
-    );
-    assert!(
-        !tasks.contains("self_type_context_validation_tasks_collect_declarations"),
-        "precollection task tests should live in precollection_tasks.rs"
-    );
-    assert!(
-        precollection.contains("self_type_context_validation_tasks_collect_declarations"),
-        "precollection_tasks.rs should cover self type context task collection"
-    );
-    assert!(
-        precollection.contains("ast_declaration_collection_bundle_replays_collection_passes"),
-        "precollection_tasks.rs should cover declaration collection task replay"
-    );
-    assert!(
-        module.contains("mod precollection_tasks;"),
-        "declaration_validation.rs should include the focused precollection_tasks module"
-    );
-}
-
-#[test]
-fn struct_literal_default_tests_live_in_focused_helper() {
-    let struct_literals = read("src/typechecker/tests/core_semantics/struct_literals.rs");
-    let defaults = read("src/typechecker/tests/core_semantics/struct_literal_defaults.rs");
-    let module = read("src/typechecker/tests/core_semantics.rs");
-
-    assert!(
-        struct_literals.lines().count() < 180,
-        "struct_literals.rs should stay focused on struct literal error cases"
-    );
-    assert!(
-        !struct_literals.contains("struct_literal_uses_default_for_omitted_field"),
-        "struct literal default tests should live in struct_literal_defaults.rs"
-    );
-    assert!(
-        defaults.contains("struct_literal_uses_default_for_omitted_field"),
-        "struct_literal_defaults.rs should cover defaulted field omission"
-    );
-    assert!(
-        defaults.contains("generic_struct_literal_uses_substituted_default_for_omitted_field"),
-        "struct_literal_defaults.rs should cover generic default substitution"
-    );
-    assert!(
-        module.contains("mod struct_literal_defaults;"),
-        "core_semantics.rs should include the focused struct_literal_defaults module"
-    );
-}
-
-#[test]
-fn generic_behavior_impl_type_arg_tests_live_in_focused_helper() {
-    let impls = read("src/typechecker/tests/generic_behaviors/impls_and_requires.rs");
-    let type_args = read("src/typechecker/tests/generic_behaviors/impl_type_args.rs");
-    let module = read("src/typechecker/tests/generic_behaviors.rs");
-
-    assert!(
-        impls.lines().count() < 180,
-        "impls_and_requires.rs should stay focused on basic impl/require behavior tests"
-    );
-    assert!(
-        !impls.contains("behavior_impl_generic_behavior_without_type_args_is_error"),
-        "generic behavior impl type-argument tests should live in impl_type_args.rs"
-    );
-    assert!(
-        type_args.contains("behavior_impl_generic_behavior_without_type_args_is_error"),
-        "impl_type_args.rs should cover missing generic behavior type arguments"
-    );
-    assert!(
-        type_args.contains("behavior_impl_generic_behavior_type_arg_bound_passes_when_satisfied"),
-        "impl_type_args.rs should cover satisfied generic behavior type-argument bounds"
-    );
-    assert!(
-        module.contains("mod impl_type_args;"),
-        "generic_behaviors.rs should include the focused impl_type_args module"
-    );
-}
